+++ src/lib/classes/conference.dart
