+++ src/lib/classes/talk.dart
